Abstract db tests
